adds Category and reworks everything to work with it Instead of having one material for each category, and then multiple lengths and heights for each material, thereby losing what could be unique information about the differing materials, we have reworked it so that each category (db) gets 1 (one) Category object, containing as many Material objects as there are materials (db) connected to that category (db) by the material_to_category translation table.
This gives us a better and more understandable translation between the database world and the java one,
as well as more freedom to vary not just the length and width of materials contained in the same
category, but all the different attributes.

This is a massive change, and while I have tested that it runs and the unit-tests have been rewritten,
I am not 100% sure everything works, especially not the BillCalculator class, as I do not understandable
what it is doing. Please look through it, and run it as well before merging the pull request.